+++ makefile.inc
@@ -1,8 +1,8 @@
 AS = i686-elf-as
 GPP = i686-elf-g++
 GCC = i686-elf-gcc
-OS_CPP = /media/sdcard/ExOS
-BUILD_TO = /mnt/boot/kernel.bin
+OS_CPP = /home/garrett/workspace/ExOS
+BUILD_TO = /run/media/garrett/EXOS/boot/kernel.bin
 CFLAGS = -c -ffreestanding -O2 -Wall -Wextra -fno-rtti -fno-exceptions -std=gnu++14 -I$(OS_CPP)/src
 OBJ_DIR = $(OS_CPP)/build
